refactor sync_user_cases into callcenter app
