+++ corehq/apps/callcenter/models.py
@@ -1,0 +1,8 @@
+# Stub models file
+from couchdbkit.ext.django.schema import Document
+
+
+class _(Document):
+    pass
+
+from .signals import *
